pyproject.toml
[tool.poetry]
name = "tap-lichess"
version = "0.0.1"
description = "`tap-lichess` is a Singer tap for lichess, built with the Meltano Singer SDK."
authors = ["HichemELB"]
keywords = [
    "ELT",
    "lichess",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
zstandard = "0.19.0"
wget = "3.2"
python-chess = "1.999"
singer-sdk = { version="^0.16.0"}
fs-s3fs = { version = "^1.1.1", optional = true}
google-cloud-storage = "2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "22.12.0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_lichess"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-lichess = 'tap_lichess.tap:Taplichess.cli'
